+++ packages/aztec.js/src/abiEncoder/bilateralSwap/index.js
@@ -1,7 +1,7 @@
 const { padLeft } = require('web3-utils');
 
 const secp256k1 = require('../../secp256k1');
-const outputCoder = require('./outputCoder');
+const outputCoder = require('../outputCoder');
 
 const bilateralSwap = {};
 bilateralSwap.outputCoder = outputCoder;
