+++ packages/aztec.js/src/abiEncoder/index.js
@@ -1,10 +1,12 @@
 const joinSplit = require('./joinSplit');
 const dividendComputation = require('./dividendComputation');
 const bilateralSwap = require('./bilateralSwap');
+const outputCoder = require('./outputCoder');
 
 
 module.exports = {
     joinSplit,
     dividendComputation,
     bilateralSwap,
+    outputCoder,
 };
