+++ packages/aztec.js/src/abiEncoder/outputCoder.js
@@ -1,8 +1,8 @@
-const { padLeft } = require('web3-utils');
+const { padLeft, sha3 } = require('web3-utils');
 const BN = require('bn.js');
 
-const bn128 = require('../../bn128');
-const secp256k1 = require('../../secp256k1');
+const bn128 = require('../bn128');
+const secp256k1 = require('../secp256k1');
 
 const outputCoder = {};
 
@@ -64,7 +64,6 @@ outputCoder.decodeNotes = (notes, isOutput) => {
 };
 
 outputCoder.decodeProofOutput = (proofOutput) => {
-    // const byteLength = parseInt(proofOutput.slice(0x00, 0x40), 16);
     const inputNotesOffset = parseInt(proofOutput.slice(0x40, 0x80), 16);
     const outputNotesOffset = parseInt(proofOutput.slice(0x80, 0xc0), 16);
     const publicOwner = `0x${proofOutput.slice(0xd8, 0x100)}`;
@@ -88,6 +87,17 @@ outputCoder.decodeProofOutputs = (proofOutputsHex) => {
         const outputOffset = parseInt(proofOutputs.slice(0x80 + (i * 0x40), 0xc0 + (i * 0x40)), 16);
         return outputCoder.decodeProofOutput(proofOutputs.slice(outputOffset * 2));
     });
+};
+
+outputCoder.getProofOutput = (proofOutputsHex, i) => {
+    const proofOutputs = proofOutputsHex.slice(2);
+    const offset = parseInt(proofOutputs.slice(0x40 + (0x40 * i), 0x80 + (0x40 * i)), 16);
+    const length = parseInt(proofOutputs.slice((offset * 2) - 0x40, (offset * 2)), 16);
+    return proofOutputs.slice((offset * 2) - 0x40, (offset * 2) + (length * 2));
+};
+
+outputCoder.hashProofOutput = (proofOutput) => {
+    return sha3(`0x${proofOutput}`);
 };
 
 outputCoder.encodeOutputNote = (note) => {
